--- Sources/UI/TEMPORARY.swift
@@ -1,10 +1,0 @@
-//
-//  TEMPORARY.swift
-//  RoverUI
-//
-//  Created by Andrew Clunis on 2018-11-22.
-//  Copyright © 2018 Rover Labs Inc. All rights reserved.
-//
-
-import Foundation
-import UIKit
